tests/wpt/tests/html/canvas/offscreen/text/2d.text.drawing.style.font-relative.spacing.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.drawing.style.font-relative.spacing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.drawing.style.font-relative.spacing</h1>
<p class="desc">Testing letter spacing and word spacing with font-relative length</p>


<script>
var t = async_test("Testing letter spacing and word spacing with font-relative length");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertSame(ctx.letterSpacing, '0px', "ctx.letterSpacing", "'0px'");
  _assertSame(ctx.wordSpacing, '0px', "ctx.wordSpacing", "'0px'");

  ctx.letterSpacing = '1EX';
  ctx.wordSpacing = '1EM';
  _assertSame(ctx.letterSpacing, '1ex', "ctx.letterSpacing", "'1ex'");
  _assertSame(ctx.wordSpacing, '1em', "ctx.wordSpacing", "'1em'");

  ctx.letterSpacing = '1ch';
  ctx.wordSpacing = '1ic';
  _assertSame(ctx.letterSpacing, '1ch', "ctx.letterSpacing", "'1ch'");
  _assertSame(ctx.wordSpacing, '1ic', "ctx.wordSpacing", "'1ic'");
  t.done();

});
</script>
