LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.font-relative.spacing-expected.txt

2d.text.drawing.style.font-relative.spacing

Testing letter spacing and word spacing with font-relative length


FAIL Testing letter spacing and word spacing with font-relative length assert_equals: ctx.letterSpacing === '0px' (got [undefined], expected 0px[string]) expected (string) "0px" but got (undefined) undefined
